requirements.txt
pandas
awswrangler
aws-glue-sessions
boto3